test(cli): split see receipt coverage
