+++ Apps/CLI/Tests/CLIAutomationTests/SeeCommandTests.swift
@@ -372,118 +372,6 @@ struct SeeCommandRuntimeTests {
 
     @Test
     @MainActor
-    func `tree only See publishes elements only with an exact action receipt`() async throws {
-        try await self.withTempConfigEnv { _ in
-            let fixture = Self.makeSeeCommandRuntimeFixture()
-            let automation = StubAutomationService()
-            automation.inspectAccessibilityTreeHandler = { _ in fixture.detectionResult }
-            let (context, _) = Self.makeSeeCommandRuntimeContext(
-                automation: automation,
-                screenCapture: fixture.screenCapture,
-                applicationInfo: fixture.applicationInfo,
-                windowInfo: fixture.windowInfo
-            )
-
-            let result = try await InProcessCommandRunner.run(
-                [
-                    "see",
-                    "--app", fixture.applicationInfo.name,
-                    "--tree",
-                    "--no-screenshot",
-                    "--json",
-                ],
-                services: context.services
-            )
-            let data = try #require(result.stdout.data(using: .utf8))
-            let response = try JSONDecoder().decode(CodableJSONResponse<SeeResult>.self, from: data)
-            let stored = try #require(
-                context.snapshots.detectionResults[response.data.snapshot_id]?.metadata.windowContext
-            )
-
-            #expect(result.exitStatus == 0)
-            #expect(response.data.ui_elements.map(\.id) == ["B1"])
-            #expect(stored.applicationProcessId == fixture.applicationInfo.processIdentifier)
-            #expect(stored.windowID == fixture.windowInfo.windowID)
-            #expect(stored.windowBounds == fixture.windowInfo.bounds)
-            #expect(stored.windowMutationIdentity == fixture.windowInfo.mutationIdentity)
-        }
-    }
-
-    @Test
-    @MainActor
-    func `tree only See refuses incomplete action receipts before snapshot publication`() async throws {
-        try await self.withTempConfigEnv { _ in
-            let fixture = Self.makeSeeCommandRuntimeFixture()
-            let fixtureContext = try #require(fixture.detectionResult.metadata.windowContext)
-            let invalidIdentities: [WindowMutationIdentity?] = [
-                nil,
-                WindowMutationIdentity(
-                    windowID: fixture.windowInfo.windowID,
-                    ownerProcessIdentifier: fixture.applicationInfo.processIdentifier,
-                    ownerProcessStartIdentity: 4242,
-                    capturedBounds: nil
-                ),
-                WindowMutationIdentity(
-                    windowID: fixture.windowInfo.windowID,
-                    ownerProcessIdentifier: fixture.applicationInfo.processIdentifier,
-                    ownerProcessStartIdentity: 0,
-                    capturedBounds: fixture.windowInfo.bounds
-                ),
-            ]
-
-            for invalidIdentity in invalidIdentities {
-                let automation = StubAutomationService()
-                automation.inspectAccessibilityTreeHandler = { _ in
-                    ElementDetectionResult(
-                        snapshotId: "incomplete-receipt-tree",
-                        screenshotPath: "",
-                        elements: fixture.detectionResult.elements,
-                        metadata: DetectionMetadata(
-                            detectionTime: 0.1,
-                            elementCount: fixture.detectionResult.elements.all.count,
-                            method: "stub",
-                            windowContext: WindowContext(
-                                applicationName: fixtureContext.applicationName,
-                                applicationBundleId: fixtureContext.applicationBundleId,
-                                applicationProcessId: fixtureContext.applicationProcessId,
-                                windowTitle: fixtureContext.windowTitle,
-                                windowID: fixtureContext.windowID,
-                                windowBounds: fixtureContext.windowBounds,
-                                windowMutationIdentity: invalidIdentity
-                            )
-                        )
-                    )
-                }
-                let (context, _) = Self.makeSeeCommandRuntimeContext(
-                    automation: automation,
-                    screenCapture: fixture.screenCapture,
-                    applicationInfo: fixture.applicationInfo,
-                    windowInfo: fixture.windowInfo
-                )
-
-                let result = try await InProcessCommandRunner.run(
-                    [
-                        "see",
-                        "--app", fixture.applicationInfo.name,
-                        "--tree",
-                        "--no-screenshot",
-                        "--json",
-                    ],
-                    services: context.services
-                )
-
-                #expect(result.exitStatus == 1)
-                #expect(result.combinedOutput.contains("exact process-generation, window, and bounds receipt"))
-                #expect(!result.combinedOutput.contains("\"snapshot_id\""))
-                #expect(!result.combinedOutput.contains("\"ui_elements\""))
-                #expect(context.snapshots.detectionResults.isEmpty)
-                #expect(try await context.snapshots.listSnapshots().isEmpty)
-            }
-        }
-    }
-
-    @Test
-    @MainActor
     func `Remote See publishes a host-certified observation without a caller barrier`() async throws {
         try await self.withTempConfigEnv { tempDir in
             let fixture = Self.makeSeeCommandRuntimeFixture()
@@ -1112,7 +1000,7 @@ struct SeeCommandRuntimeTests {
         }
     }
 
-    private func withTempConfigEnv<T>(
+    func withTempConfigEnv<T>(
         _ body: @escaping (URL) async throws -> T
     ) async throws -> T {
         let tempDir = FileManager.default.temporaryDirectory.appendingPathComponent(
@@ -1143,7 +1031,7 @@ struct SeeCommandRuntimeTests {
 }
 
 extension SeeCommandRuntimeTests {
-    fileprivate struct RuntimeFixture {
+    struct RuntimeFixture {
         let snapshotId: String
         let applicationInfo: ServiceApplicationInfo
         let windowInfo: ServiceWindowInfo
@@ -1151,7 +1039,7 @@ extension SeeCommandRuntimeTests {
         let detectionResult: ElementDetectionResult
     }
 
-    fileprivate static func makeSeeCommandRuntimeFixture() -> RuntimeFixture {
+    static func makeSeeCommandRuntimeFixture() -> RuntimeFixture {
         let snapshotId = UUID().uuidString
         let windowBounds = CGRect(x: 10, y: 20, width: 800, height: 600)
         let applicationInfo = Self.makeSeeFixtureApplicationInfo()
@@ -1177,7 +1065,7 @@ extension SeeCommandRuntimeTests {
         )
     }
 
-    fileprivate static func makeSeeCommandRuntimeContext(
+    static func makeSeeCommandRuntimeContext(
         automation: StubAutomationService,
         screenCapture: StubScreenCaptureService,
         applicationInfo: ServiceApplicationInfo? = nil,
